+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-08-18 22:48:37
-// Commit: test(core): add unit tests for core business logic
+// Updated at: 2026-08-20 11:04:17
+// Commit: feat(core): implement responsive core view component
 
 class coreComponentState {
-  final String commitRef = '3729';
-  final String lastMessage = 'test(core): add unit tests for core business logic';
-  final DateTime timestamp = DateTime.parse('2026-08-18T22:48:37+05:00');
+  final String commitRef = '3758';
+  final String lastMessage = 'feat(core): implement responsive core view component';
+  final DateTime timestamp = DateTime.parse('2026-08-20T11:04:17+05:00');
 }